キャラクター成長を指数関数化し、レベルアップ演出とEXP常時可視化を追加
HP/SP/ATK/DEFの成長曲線が平坦すぎる問題と、レベルアップ時のフィードバックが
チャット1行だけで演出に乏しい問題に対応した。

- config.yml: status.growth のHP/SP/ATK/DEFを、モンスターのレベルスケーリング
  (stat-scaling.growth-rate)と同じ指数成長 scaled = base * growthRate^(level-1)
  に統一。CRT/CRT_DMG/SPDは従来通り線形のまま。レベル上限は100→80(データ上300まで
  対応)。数値の桁が増えるため、モンスター名札のHPバー表示もintからlongに変更。
- LevelUpFeedbackServiceを新設し、タイトル・サウンド・パーティクル
  (status.level-up-effectで設定、無効な名前は黙ってスキップ)と、伸びたステータス
  だけを表示するチャット行を追加。ジョブ/採取レベルのレベルアップ
  (GatheringLevelService)も同じサービス経由に統一。
- アクションバーに{level}/{exp_bar}(次レベルまでの進捗バー)、ステータスGUIの
  頭アイコンにも経験値(現在値/必要値、上限到達時はMAX)を常時表示。

+++ src/main/java/rpg/monster/service/MonsterHealthBarRenderer.java
@@ -28,7 +28,7 @@ public final class MonsterHealthBarRenderer {
         return format
                 .replace("{name}", name)
                 .replace("{bar}", bar)
-                .replace("{current}", String.valueOf((int) Math.ceil(Math.max(0, currentHp))))
-                .replace("{max}", String.valueOf((int) Math.ceil(maxHp)));
+                .replace("{current}", String.valueOf((long) Math.ceil(Math.max(0, currentHp))))
+                .replace("{max}", String.valueOf((long) Math.ceil(maxHp)));
     }
 }
